@@ -3,11 +3,11 @@
 namespace App;
 
 use DateTime;
-use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Database\Eloquent\Relations\HasMany;
 
 /**
  * @property string|null ip
@@ -137,7 +137,7 @@ class CustomerBack extends Model
     /** @var string */
     protected $table = 'SS_buyers_gamepost';
 
-    /** @var array<string, mixed> */
+    /** @psalm-var array<string, mixed> */
     protected $attributes = [
         self::vip => '',
         self::info => '',
