@@ -5,11 +5,11 @@ namespace App\Events;
 class EventList
 {
     /** @var string */
-    const ORDER_NEW = 'order_new';
+    const TEST = 'test';
 
     /** @var string */
     const BACKUP = 'backup';
 
     /** @var string */
-    const TEST = 'test';
+    const ORDER_NEW = 'order_new';
 }
