@@ -3,43 +3,12 @@
 namespace App\Exceptions;
 
 use Illuminate\Foundation\Exceptions\Handler as ExceptionHandler;
-use Illuminate\Http\Request;
-use Symfony\Component\HttpFoundation\Response;
-use Throwable;
 
 class Handler extends ExceptionHandler
 {
-    /**
-     * @var string[]
-     */
-    protected $dontReport = [];
-
-    /**
-     * @var string[]
-     */
+    /** @var string[] */
     protected $dontFlash = [
         'password',
         'password_confirmation',
     ];
-
-    /**
-     * @param Throwable $exception
-     * @return void
-     * @throws Throwable
-     */
-    public function report(Throwable $exception): void
-    {
-        parent::report($exception);
-    }
-
-    /**
-     * @param  Request $request
-     * @param  Throwable $exception
-     * @return Response
-     * @throws Throwable
-     */
-    public function render($request, Throwable $exception): Response
-    {
-        return parent::render($request, $exception);
-    }
 }
