@@ -1,8 +1,6 @@
 <?php
 
-$uri = urldecode(
-    parse_url($_SERVER['REQUEST_URI'], PHP_URL_PATH)
-);
+$uri = urldecode(parse_url($_SERVER['REQUEST_URI'], PHP_URL_PATH));
 
 if ('/' !== $uri && true === file_exists(__DIR__ . '/public' . $uri)) {
     return false;
